feat(app): add express server
Adds an express server to the app.  This allows the app
to be deployed to a node environment and serve the SvelteKit
app.  The express server also adds a healthcheck endpoint
at `/healthcheck`.  The server listens on port 3000.  The
SvelteKit adapter is changed to `adapter-node` to support
this.  Express is added as a dependency.